rename testproviders to testProviders & move RelativeTime component
