+++ code/core/src/manager-api/root.tsx
@@ -50,7 +50,7 @@ import { noArrayMerge } from './lib/merge';
 import type { ModuleFn } from './lib/types';
 import * as addons from './modules/addons';
 import * as channel from './modules/channel';
-import * as testproviders from './modules/experimental_testmodule';
+import * as testProviders from './modules/experimental_testmodule';
 import * as globals from './modules/globals';
 import * as layout from './modules/layout';
 import * as notifications from './modules/notifications';
@@ -80,7 +80,7 @@ export type State = layout.SubState &
   stories.SubState &
   refs.SubState &
   notifications.SubState &
-  testproviders.SubState &
+  testProviders.SubState &
   version.SubState &
   url.SubState &
   shortcuts.SubState &
@@ -100,7 +100,7 @@ export type API = addons.SubAPI &
   globals.SubAPI &
   layout.SubAPI &
   notifications.SubAPI &
-  testproviders.SubAPI &
+  testProviders.SubAPI &
   shortcuts.SubAPI &
   settings.SubAPI &
   version.SubAPI &
@@ -181,7 +181,7 @@ class ManagerProvider extends Component<ManagerProviderProps, State> {
       addons,
       layout,
       notifications,
-      testproviders,
+      testProviders,
       settings,
       shortcuts,
       stories,
